+++ src/webviews/src/modules/key-details/components/rejson-details/rejson-object/RejsonObject.tsx
@@ -5,8 +5,8 @@ import { diff } from 'deep-diff'
 import { useShallow } from 'zustand/react/shallow'
 import { Spinner } from 'uiSrc/ui'
 
-import { checkExistingPath } from './tbd'
 import ReJSONConfirmDialog from './RejsonConfirmDialog'
+import { checkExistingPath } from '../utils/path'
 import { RejsonDynamicTypes } from '../rejson-dynamic-types'
 import { JSONObjectProps, ObjectTypes } from '../interfaces'
 import { generatePath, getBrackets, wrapPath } from '../utils'
